# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "Cleared": {
        "properties": {
          "cleared": {
            "title": "Cleared",
            "type": "integer"
          }
        },
        "required": [
          "cleared"
        ],
        "title": "Cleared",
        "type": "object"
      },
      "Device": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "lite",
              "lite_flatport",
              "mobile",
              "multilens",
              "mono",
              "scout"
            ],
            "title": "Kind",
            "type": "string"
          },
          "serial": {
            "title": "Serial",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "serial"
        ],
        "title": "Device",
        "type": "object"
      },
      "DeviceCreate": {
        "properties": {
          "kind": {
            "enum": [
              "lite",
              "lite_flatport",
              "mobile",
              "multilens",
              "mono",
              "scout"
            ],
            "title": "Kind",
            "type": "string"
          },
          "serial": {
            "title": "Serial",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "serial"
        ],
        "title": "DeviceCreate",
        "type": "object"
      },
      "Dive": {
        "properties": {
          "calibration_source_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Source Number"
          },
          "calibration_target_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Target Number"
          },
          "dived_at": {
            "format": "date-time",
            "title": "Dived At",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "priority": {
            "enum": [
              "low",
              "high",
              "none"
            ],
            "title": "Priority",
            "type": "string"
          },
          "slate_template_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slate Template Number"
          }
        },
        "required": [
          "number",
          "name",
          "dived_at",
          "priority",
          "slate_template_number",
          "calibration_target_number",
          "calibration_source_number"
        ],
        "title": "Dive",
        "type": "object"
      },
      "Flagged": {
        "properties": {
          "flagged": {
            "title": "Flagged",
            "type": "integer"
          }
        },
        "required": [
          "flagged"
        ],
        "title": "Flagged",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MyMembership": {
        "properties": {
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "is_admin"
        ],
        "title": "MyMembership",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FishSense Services API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/tenants/{slug}/devices": {
      "get": {
        "operationId": "list_devices",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  },
                  "title": "Response List Devices",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Devices"
      },
      "post": {
        "operationId": "create_device",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Device"
      }
    },
    "/tenants/{slug}/dives": {
      "get": {
        "operationId": "list_dives",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Dive"
                  },
                  "title": "Response List Dives",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dives"
      }
    },
    "/tenants/{slug}/dives/{number}/calibration-refusal": {
      "delete": {
        "description": "Clear dive ``number``'s standing calibration refusal, so it is\nfitted again (idempotent): for a change its labels don't show.",
        "operationId": "clear_dive_calibration_refusal",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "why, kept with the clear",
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "why, kept with the clear",
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Dive Calibration Refusal"
      }
    },
    "/tenants/{slug}/dives/{number}/calibration-source": {
      "delete": {
        "description": "Unlink dive ``number`` from any borrowed calibration (idempotent).",
        "operationId": "clear_dive_calibration_source",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Dive Calibration Source"
      }
    },
    "/tenants/{slug}/dives/{number}/calibration-source/{source_number}": {
      "put": {
        "description": "Dive ``number`` borrows dive ``source_number``'s laser calibration.\n\n400 on a self-link; 404 if either dive is missing from the tenant.",
        "operationId": "set_dive_calibration_source",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "source_number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Source Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dive"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Dive Calibration Source"
      }
    },
    "/tenants/{slug}/dives/{number}/calibration-target": {
      "delete": {
        "description": "Unlink dive ``number`` from any calibration target (idempotent): it\nleaves the checkerboard cohort. A standing refusal is left as it is.",
        "operationId": "clear_dive_calibration_target",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Dive Calibration Target"
      }
    },
    "/tenants/{slug}/dives/{number}/calibration-target/{target_number}": {
      "put": {
        "description": "Dive ``number`` was shot against calibration target\n``target_number``: it enters the checkerboard cohort, and a standing\ncalibration refusal expires.\n\n404 if the dive is missing from the tenant or no target has that\nnumber.",
        "operationId": "set_dive_calibration_target",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "target_number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Target Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dive"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Dive Calibration Target"
      }
    },
    "/tenants/{slug}/dives/{number}/labels/{kind}/needs-reprocess": {
      "delete": {
        "description": "Withdraw a redraw of dive ``number``'s ``kind`` frames (idempotent:\n0 when none was asked for).",
        "operationId": "clear_needs_reprocess",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "laser",
                "head_tail",
                "species",
                "slate"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cleared"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Needs Reprocess"
      },
      "put": {
        "description": "Redraw dive ``number``'s ``kind`` frames: the dive re-enters that\nkind's preprocessing, and the JPEGs are redrawn where they are.\nIncomplete labels only unless ``only_incomplete=false``.",
        "operationId": "raise_needs_reprocess",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Number",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "laser",
                "head_tail",
                "species",
                "slate"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "only_incomplete",
            "required": false,
            "schema": {
              "default": true,
              "title": "Only Incomplete",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Flagged"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Raise Needs Reprocess"
      }
    },
    "/tenants/{slug}/label-studio-projects": {
      "get": {
        "description": "Label Studio project ids with live labels of ``kind``.",
        "operationId": "list_label_studio_project_ids",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "laser",
                "head_tail",
                "species",
                "slate"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "incomplete",
            "required": false,
            "schema": {
              "default": false,
              "title": "Incomplete",
              "type": "boolean"
            }
          },
          {
            "description": "laser only: the auto-accept gate is done here",
            "in": "query",
            "name": "gated",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "laser only: the auto-accept gate is done here",
              "title": "Gated"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "integer"
                  },
                  "title": "Response List Label Studio Project Ids",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Label Studio Project Ids"
      }
    },
    "/tenants/{slug}/membership": {
      "get": {
        "operationId": "get_my_membership",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMembership"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Membership"
      }
    }
  }
}